+++ cpskin/core/viewlets/organization_gallery.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from collective.contact.core.content.organization import IOrganization
+from collective.folderishtypes.interfaces import IFolderishType
 from plone import api
 from plone.app.layout.viewlets import common
 
@@ -8,6 +10,15 @@ from plone.app.layout.viewlets import common
 class OrganizationGalleryViewlet(common.ViewletBase):
 
     index = ViewPageTemplateFile('organization_gallery.pt')
+
+    def available(self):
+        """
+        Condition on the content types wanted for the gallery viewlet
+        (it was first meant only for Organizations), now it is also used for
+        folderish content types (collective.folderish)
+        """
+        context = self.context
+        return IOrganization.providedBy(context) or IFolderishType.providedBy(context)
 
     def get_photos(self):
         brains = api.content.find(